# openAPI/generated-source/src/main/java/com/example/samplesystem/sample/model/HelloWorldResponse.java
/*
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.example.samplesystem.sample.model;

import java.util.Objects;
import java.util.Arrays;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import com.fasterxml.jackson.annotation.JsonTypeName;

/**
 * HelloWorldResponse
 */
@JsonPropertyOrder({
  HelloWorldResponse.JSON_PROPERTY_HELLO
})
@JsonTypeName("HelloWorldResponse")
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2021-12-07T09:02:24.157159Z[Etc/UTC]")
public class HelloWorldResponse {
  public static final String JSON_PROPERTY_HELLO = "hello";
  private String hello;


  public HelloWorldResponse hello(String hello) {
    
    this.hello = hello;
    return this;
  }

   /**
   * Get hello
   * @return hello
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")
  @JsonProperty(JSON_PROPERTY_HELLO)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getHello() {
    return hello;
  }


  @JsonProperty(JSON_PROPERTY_HELLO)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setHello(String hello) {
    this.hello = hello;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    HelloWorldResponse helloWorldResponse = (HelloWorldResponse) o;
    return Objects.equals(this.hello, helloWorldResponse.hello);
  }

  @Override
  public int hashCode() {
    return Objects.hash(hello);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class HelloWorldResponse {\n");
    sb.append("    hello: ").append(toIndentedString(hello)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}