# openAPI/generated-source/src/main/java/com/example/samplesystem/sample/model/Ytorderadd.java
/*
 * OpenAPI definition
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


package com.example.samplesystem.sample.model;

import java.util.Objects;
import java.util.Arrays;
import com.fasterxml.jackson.annotation.JsonInclude;
import com.fasterxml.jackson.annotation.JsonProperty;
import com.fasterxml.jackson.annotation.JsonCreator;
import com.fasterxml.jackson.annotation.JsonTypeName;
import com.fasterxml.jackson.annotation.JsonValue;
import io.swagger.annotations.ApiModel;
import io.swagger.annotations.ApiModelProperty;
import com.fasterxml.jackson.annotation.JsonPropertyOrder;
import com.fasterxml.jackson.annotation.JsonTypeName;

/**
 * Ytorderadd
 */
@JsonPropertyOrder({
  Ytorderadd.JSON_PROPERTY_MANDT,
  Ytorderadd.JSON_PROPERTY_YBUKRS,
  Ytorderadd.JSON_PROPERTY_YBELN,
  Ytorderadd.JSON_PROPERTY_YREASONCD,
  Ytorderadd.JSON_PROPERTY_YREASONS
})
@JsonTypeName("Ytorderadd")
@javax.annotation.Generated(value = "org.openapitools.codegen.languages.JavaClientCodegen", date = "2021-12-07T09:02:24.157159Z[Etc/UTC]")
public class Ytorderadd {
  public static final String JSON_PROPERTY_MANDT = "mandt";
  private String mandt;

  public static final String JSON_PROPERTY_YBUKRS = "ybukrs";
  private String ybukrs;

  public static final String JSON_PROPERTY_YBELN = "ybeln";
  private String ybeln;

  public static final String JSON_PROPERTY_YREASONCD = "yreasoncd";
  private String yreasoncd;

  public static final String JSON_PROPERTY_YREASONS = "yreasons";
  private String yreasons;


  public Ytorderadd mandt(String mandt) {
    
    this.mandt = mandt;
    return this;
  }

   /**
   * Get mandt
   * @return mandt
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")
  @JsonProperty(JSON_PROPERTY_MANDT)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getMandt() {
    return mandt;
  }


  @JsonProperty(JSON_PROPERTY_MANDT)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setMandt(String mandt) {
    this.mandt = mandt;
  }


  public Ytorderadd ybukrs(String ybukrs) {
    
    this.ybukrs = ybukrs;
    return this;
  }

   /**
   * Get ybukrs
   * @return ybukrs
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")
  @JsonProperty(JSON_PROPERTY_YBUKRS)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getYbukrs() {
    return ybukrs;
  }


  @JsonProperty(JSON_PROPERTY_YBUKRS)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setYbukrs(String ybukrs) {
    this.ybukrs = ybukrs;
  }


  public Ytorderadd ybeln(String ybeln) {
    
    this.ybeln = ybeln;
    return this;
  }

   /**
   * Get ybeln
   * @return ybeln
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")
  @JsonProperty(JSON_PROPERTY_YBELN)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getYbeln() {
    return ybeln;
  }


  @JsonProperty(JSON_PROPERTY_YBELN)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setYbeln(String ybeln) {
    this.ybeln = ybeln;
  }


  public Ytorderadd yreasoncd(String yreasoncd) {
    
    this.yreasoncd = yreasoncd;
    return this;
  }

   /**
   * Get yreasoncd
   * @return yreasoncd
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")
  @JsonProperty(JSON_PROPERTY_YREASONCD)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getYreasoncd() {
    return yreasoncd;
  }


  @JsonProperty(JSON_PROPERTY_YREASONCD)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setYreasoncd(String yreasoncd) {
    this.yreasoncd = yreasoncd;
  }


  public Ytorderadd yreasons(String yreasons) {
    
    this.yreasons = yreasons;
    return this;
  }

   /**
   * Get yreasons
   * @return yreasons
  **/
  @javax.annotation.Nullable
  @ApiModelProperty(value = "")
  @JsonProperty(JSON_PROPERTY_YREASONS)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)

  public String getYreasons() {
    return yreasons;
  }


  @JsonProperty(JSON_PROPERTY_YREASONS)
  @JsonInclude(value = JsonInclude.Include.USE_DEFAULTS)
  public void setYreasons(String yreasons) {
    this.yreasons = yreasons;
  }


  @Override
  public boolean equals(Object o) {
    if (this == o) {
      return true;
    }
    if (o == null || getClass() != o.getClass()) {
      return false;
    }
    Ytorderadd ytorderadd = (Ytorderadd) o;
    return Objects.equals(this.mandt, ytorderadd.mandt) &&
        Objects.equals(this.ybukrs, ytorderadd.ybukrs) &&
        Objects.equals(this.ybeln, ytorderadd.ybeln) &&
        Objects.equals(this.yreasoncd, ytorderadd.yreasoncd) &&
        Objects.equals(this.yreasons, ytorderadd.yreasons);
  }

  @Override
  public int hashCode() {
    return Objects.hash(mandt, ybukrs, ybeln, yreasoncd, yreasons);
  }

  @Override
  public String toString() {
    StringBuilder sb = new StringBuilder();
    sb.append("class Ytorderadd {\n");
    sb.append("    mandt: ").append(toIndentedString(mandt)).append("\n");
    sb.append("    ybukrs: ").append(toIndentedString(ybukrs)).append("\n");
    sb.append("    ybeln: ").append(toIndentedString(ybeln)).append("\n");
    sb.append("    yreasoncd: ").append(toIndentedString(yreasoncd)).append("\n");
    sb.append("    yreasons: ").append(toIndentedString(yreasons)).append("\n");
    sb.append("}");
    return sb.toString();
  }

  /**
   * Convert the given object to string with each line indented by 4 spaces
   * (except the first line).
   */
  private String toIndentedString(Object o) {
    if (o == null) {
      return "null";
    }
    return o.toString().replace("\n", "\n    ");
  }

}